подключение к базе, сущности

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -5,6 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.skypro.homework.entity.Ad;
 
 @Repository
-public interface AdRepository extends JpaRepository<Ad,Long> {
-
+public interface AdRepository extends JpaRepository<Ad, Integer> {
 }
