+++ src/main/java/ru/skypro/homework/repository/ImageRepository.java
@@ -5,6 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.skypro.homework.entity.Image;
 
 @Repository
-public interface ImageRepository  extends JpaRepository<Image, Long> {
-
+public interface ImageRepository  extends JpaRepository<Image, Integer> {
 }
