+++ src/main/java/ru/skypro/homework/repository/UserRepository.java
@@ -5,6 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.skypro.homework.entity.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
